+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 1.0.3
 * added travis-ci integration
+* added serverspec tests
 
 # 1.0.2
 * switched to mongodb 2.6, thus fixed ubuntu installation